Master

+++ app/Http/Controllers/AdminVideosController.php
@@ -1187,7 +1187,7 @@ class AdminVideosController extends Controller
         }
 
         $data['video_js_pre_position_ads'] = $request->video_js_pre_position_ads ;
-        $data['video_js_post_position_ads'] = $request->video_js_post_position_ads ;
+        $data['video_js_post_position_ads'] = $request->video_js_pre_position_ads ;
         $data['video_js_mid_position_ads_category'] = $request->video_js_mid_position_ads_category ;
         $data['video_js_mid_advertisement_sequence_time'] = $request->video_js_mid_advertisement_sequence_time ;
         $data['expiry_date'] = $request->expiry_date ;
@@ -1411,16 +1411,20 @@ class AdminVideosController extends Controller
             }
 
                     //  Delete Existing Trailer Video - M3u8 Format
-            $video_trailer_m3u8 = pathinfo($videos->trailer)['filename'];
 
-            $directory = base_path('public/uploads/trailer/');
-                    
-            $pattern =  $video_trailer_m3u8.'*';
+            if (!is_null($videos->trailer)) {
 
-            $files = glob($directory . $pattern);
+                $video_trailer_m3u8 = pathinfo($videos->trailer)['filename'];
 
-            foreach ($files as $file) {
-                File::delete($file);
+                $directory = base_path('public/uploads/trailer/');
+                        
+                $pattern =  $video_trailer_m3u8.'*';
+
+                $files = glob($directory . $pattern);
+
+                foreach ($files as $file) {
+                    File::delete($file);
+                }
             }
 
                      //  Delete Existing Trailer Video - MP4 Format
@@ -1433,14 +1437,20 @@ class AdminVideosController extends Controller
                     //  Delete Existing  Video
             $directory = storage_path('app/public');
                     
-            $info = pathinfo($videos->path);
+            if (!is_null($videos->path)) {
 
-            $pattern =  $info['filename'] . '*';
+                $info = pathinfo($videos->path);
 
-            $files = glob($directory . '/' . $pattern);
-
-            foreach ($files as $file) {
-                unlink($file);
+                $pattern =  $videos->path ? $info['filename'] . '*' : " ";
+    
+                $files = glob($directory . '/' . $pattern);
+    
+                foreach ($files as $file) {
+    
+                    if(file_exists( $file )){
+                        unlink($file);
+                    }
+                }   
             }
 
                     // Video uploaded by CPP user while deleting Mail
@@ -1658,7 +1668,6 @@ class AdminVideosController extends Controller
         $settings = Setting::first();
 
         $video = Video::findOrFail($id);
-
         Video::query()->where('id','!=', $id)->update(['today_top_video' => 0]);
 
         if ($request->slug == "") {
@@ -1737,7 +1746,9 @@ class AdminVideosController extends Controller
         // Trailer Update
 
         $path = public_path() . "/uploads/videos/";
-        $video->trailer_type = $data["trailer_type"];
+        if($trailer != ""){
+            $video->trailer_type = $data["trailer_type"];
+        }
 
         $StorageSetting = StorageSetting::first();
 
@@ -1748,6 +1759,7 @@ class AdminVideosController extends Controller
                 $settings->transcoding_access == 1 &&
                 $data["trailer_type"] == "video_mp4"
             ) {
+
                 if ($settings->transcoding_resolution != null) {
                     $convertresolution = [];
                     $resolution = explode(",", $settings->transcoding_resolution);
@@ -1776,27 +1788,28 @@ class AdminVideosController extends Controller
                         }
                         if ($value == "1080p") {
                             $r_1080p = (new Representation())
-                                ->setKiloBitrate(750)
-                                ->setResize(854, 480);
+                                ->setKiloBitrate(4096)
+                                ->setResize(1920, 1080);
                             array_push($convertresolution, $r_1080p);
                         }
                     }
                 }
                 $trailer = $data["trailer"];
-                $trailer_path = URL::to("public/uploads/trailer/");
+                $trailer_path = URL::to("storage/app/trailer/");
                 $trailer_Videoname =  Str::lower($trailer->getClientOriginalName());
                 $trailer_Video = time() . "_" . str_replace(" ","_",$trailer_Videoname);
-                $trailer->move(public_path("uploads/trailer/"), $trailer_Video);
+                $trailer->move(storage_path("app/trailer/"), $trailer_Video);
                 $trailer_video_name = strtok($trailer_Video, ".");
                 $M3u8_save_path =
                     $trailer_path . "/" . $trailer_video_name . ".m3u8";
-                $storepath = URL::to("public/uploads/trailer/");
+                $storepath = URL::to("storage/app/trailer/");
     
                 $data["trailer"] = $M3u8_save_path;
                 $video->trailer_type = "m3u8";
                 $data["trailer_type"] = "m3u8";
             } else {
-                if ($data["trailer_type"] == "video_mp4") {
+                
+                if ($trailer != "" && $data["trailer_type"] == "video_mp4") {
                     if (!empty($trailer)) {
                         if ($trailer != "" && $trailer != null) {
                             $file_old = $path . $trailer;
@@ -1818,6 +1831,7 @@ class AdminVideosController extends Controller
                             URL::to("/") . "/public/uploads/videos/" . $trailer_vid;
                     } else {
                         $data["trailer"] = $video->trailer;
+                        $data["trailer_type"] = $video->trailer_type;
                     }
                 } elseif ($data["trailer_type"] == "m3u8_url") {
                     $video->trailer = $data["m3u8_trailer"];
@@ -1830,6 +1844,9 @@ class AdminVideosController extends Controller
                     $data["trailer"] = $data["embed_trailer"];
                 } else {
                     $data["trailer"] = $video->trailer;
+                    $data["trailer_type"] = $video->trailer_type;
+                    // dd('test'.$video);
+
                 }
                 // $data['trailer'] = "";
             }
@@ -1907,27 +1924,27 @@ class AdminVideosController extends Controller
                         }
                         if ($value == "1080p") {
                             $r_1080p = (new Representation())
-                                ->setKiloBitrate(750)
-                                ->setResize(854, 480);
+                                ->setKiloBitrate(4096)
+                                ->setResize(1920, 1080);
                             array_push($convertresolution, $r_1080p);
                         }
                     }
                 }
                 $trailer = $data["trailer"];
-                $trailer_path = URL::to("public/uploads/trailer/");
+                $trailer_path = URL::to("storage/app/trailer/");
                 $trailer_Videoname =  Str::lower($trailer->getClientOriginalName());
                 $trailer_Video = time() . "_" . str_replace(" ","_",$trailer_Videoname);
-                $trailer->move(public_path("uploads/trailer/"), $trailer_Video);
+                $trailer->move(storage_path("app/trailer/"), $trailer_Video);
                 $trailer_video_name = strtok($trailer_Video, ".");
                 $M3u8_save_path =
                     $trailer_path . "/" . $trailer_video_name . ".m3u8";
-                $storepath = URL::to("public/uploads/trailer/");
+                $storepath = URL::to("storage/app/trailer/");
     
                 $data["trailer"] = $M3u8_save_path;
                 $video->trailer_type = "m3u8";
                 $data["trailer_type"] = "m3u8";
             } else {
-                if ($data["trailer_type"] == "video_mp4") {
+                if ( $trailer != "" && $data["trailer_type"] == "video_mp4") {
                     if (!empty($trailer)) {
                         if ($trailer != "" && $trailer != null) {
                             $file_old = $path . $trailer;
@@ -1949,6 +1966,8 @@ class AdminVideosController extends Controller
                             URL::to("/") . "/public/uploads/videos/" . $trailer_vid;
                     } else {
                         $data["trailer"] = $video->trailer;
+                        $data["trailer_type"] = $video->trailer_type;
+
                     }
                 } elseif ($data["trailer_type"] == "m3u8_url") {
                     $video->trailer = $data["m3u8_trailer"];
@@ -1961,8 +1980,9 @@ class AdminVideosController extends Controller
                     $data["trailer"] = $data["embed_trailer"];
                 } else {
                     $data["trailer"] = $video->trailer;
+                    $data["trailer_type"] = $video->trailer_type;
                 }
-            }
+                }
             }
 
 
@@ -2530,7 +2550,7 @@ class AdminVideosController extends Controller
         $video->search_tags = $searchtags;
         $video->ios_ppv_price = $request->ios_ppv_price;
         $video->video_js_pre_position_ads = $request->video_js_pre_position_ads;
-        $video->video_js_post_position_ads = $request->video_js_post_position_ads;
+        $video->video_js_post_position_ads = $request->video_js_pre_position_ads;
         $video->video_js_mid_position_ads_category = $request->video_js_mid_position_ads_category;
         $video->video_js_mid_advertisement_sequence_time = $request->video_js_mid_advertisement_sequence_time;
         $video->expiry_date = $request->expiry_date;
@@ -3311,22 +3331,22 @@ class AdminVideosController extends Controller
                                 array_push($convertresolution, $r_720p);
                             }
                             if ($value == '1080p') {
-                                $r_1080p = (new Representation())->setKiloBitrate(750)->setResize(854, 480);
+                                $r_1080p = (new Representation())->setKiloBitrate(4096)->setResize(1920, 1080);
                                 array_push($convertresolution, $r_1080p);
                             }
                         }
                     }
                     $trailer = $data['trailer'];
-                    $trailer_path = URL::to('public/uploads/trailer/');
+                    $trailer_path = URL::to('storage/app/trailer/');
                     $trailer_Videoname = Str::lower($trailer->getClientOriginalName());
                     $trailer_Video = time() . '_' . str_replace(' ', '_', $trailer_Videoname);
 
                     // $trailer_Video =
                     //     time() . "_" . $trailer->getClientOriginalName();
-                    $trailer->move(public_path('uploads/trailer/'), $trailer_Video);
+                    $trailer->move(storage_path('app/trailer/'), $trailer_Video);
                     $trailer_video_name = strtok($trailer_Video, '.');
                     $M3u8_save_path = $trailer_path . '/' . $trailer_video_name . '.m3u8';
-                    $storepath = URL::to('public/uploads/trailer/');
+                    $storepath = URL::to('storage/app/trailer/');
 
                     $data['trailer'] = $M3u8_save_path;
                     $data['trailer_type'] = 'm3u8';
@@ -3413,21 +3433,21 @@ class AdminVideosController extends Controller
                                 array_push($convertresolution, $r_720p);
                             }
                             if ($value == '1080p') {
-                                $r_1080p = (new Representation())->setKiloBitrate(750)->setResize(854, 480);
+                                $r_1080p = (new Representation())->setKiloBitrate(4096)->setResize(1920, 1080);
                                 array_push($convertresolution, $r_1080p);
                             }
                         }
                     }
                     $trailer = $data['trailer'];
-                    $trailer_path = URL::to('public/uploads/trailer/');
+                    $trailer_path = URL::to('storage/app/trailer/');
                     $trailer_Videoname = Str::lower($trailer->getClientOriginalName());
                     $trailer_Video = time() . '_' . str_replace(' ', '_', $trailer_Videoname);
                     // $trailer_Video =
                     //     time() . "_" . $trailer->getClientOriginalName();
-                    $trailer->move(public_path('uploads/trailer/'), $trailer_Video);
+                    $trailer->move(storage_path('app/trailer/'), $trailer_Video);
                     $trailer_video_name = strtok($trailer_Video, '.');
                     $M3u8_save_path = $trailer_path . '/' . $trailer_video_name . '.m3u8';
-                    $storepath = URL::to('public/uploads/trailer/');
+                    $storepath = URL::to('storage/app/trailer/');
 
                     $data['trailer'] = $M3u8_save_path;
                     $data['trailer_type'] = 'm3u8';
